+++ framy_generator/lib/resolvers/widget_resolver.dart
@@ -80,6 +80,9 @@ class WidgetResolver extends GeneratorForAnnotation<FramyWidget> {
           ),
         );
       }
+
+      //add groupName
+      framyObject.widgetGroupName = annotation.peek('groupName')?.stringValue;
     }
 
     framyObjectsToReturn.add(framyObject);